[FEATURE] Using new generating Identifiers for CPU model (also cpu_architecture and ram_type entities were updated)

+++ src/main/kotlin/ru/viklover/pcassembly/ram/RamRepository.kt
@@ -4,4 +4,4 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface RamRepository : CrudRepository<Ram, Long>
+interface RamRepository : CrudRepository<Ram, Int>
